Handle empty json file in Fetch method
In the case where the collector extracts json files from a remote tool before processing them, it may happen that the json file related to a given class of objects ($this->aJson) is empty. In such case, the Fetch method must take the possibility into account. This is not the case today, hence the proposal.